+++ Assessment3/lib/dac.h
@@ -7,6 +7,6 @@
 #define MAX_VOLTAGE 3.3
 
 void dac_init();
-void dac_update_data(uint32_t val);
+void dac_update_data(uint16_t val);
 void dac_update_voltage(float val);
 void dac_pincfg();
